Merging data processing with master

See merge request lab2324_summer/armv8_9!7

+++ src/memory.h
@@ -9,9 +9,9 @@
 
 typedef uint8_t BYTE;
 typedef int32_t INST;
-#define regCount 31
-#define memSize 1 << 21
-extern const int64_t HALT;
+#define regCount (31)
+#define memSize (1 << 21)
+extern const int32_t HALT;
 extern int64_t registers[regCount];
 extern int64_t programCounter;
 extern int64_t zeroRegister;
